fix: move type.ts
